+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - Changed `applyContainerStyles()` to take a single `darkmode` value instead of "options", with supported values: `true`, `false`, `'true'`, `'false'`, `'dark'`, `'light'` and `'auto'`
+- Changed `applyContainerStyles()` to avoid applying transition styles to the container, preventing unwanted flashes of opposite colors
 - Changed `source` view to take `source`, `syntax`, `lineNum`, `refs` and `maxSourceSizeToHighlight` options. It attempts to derive these values from `data` when options are not explicitly provided for backward compatibility. The `source` can be derived from `data` as `content` (for backward compatibility) or `source` property
 - Removed `mime`, `binary` and `size` data options for `source` view
 - Bumped `marked` to `^14.1.4` (used in `markdown` view)